fix: change debit/credit filter works
...we don't have UI for adding a card type yet in the 'Add new card'
form, so until then any card that has 'debit' in its title will be
considered a debit card and otherwise it's a credit card.

Once the UI is implemented, and edit-card functionality and all cards
are migrated to correct `CardType`, then we'll use the `CardType`s.
Until then, this is a hacky fix.